requirements.txt
black==23.7.0
boto3==1.28.50
configparser==6.0.0
Django==4.2.3
django-cors-headers==4.2.0
django-multiselectfield==0.1.12
django-storages==1.14
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.4
isort==5.12.0
pillow==10.0.1
whitenoise==6.5.0
